backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
openpyxl==3.1.2
pillow==10.1.0
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
boto3==1.34.0
transformers==4.36.0
torch==2.1.1
networkx==3.2.1
shap==0.44.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.0.3
python-docx==1.1.0
reportlab==4.0.7
httpx==0.25.2
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
python-json-logger==2.0.7
